+++ web/app/scripts/compiled-es5/form/templates.js
@@ -12,12 +12,12 @@ angular.module('incuna-surveys-form.templates', []).run(['$templateCache', funct
 
 
   $templateCache.put('templates/incuna-surveys/form/base/proportion-field.html',
-    "<h4 class=title ng-bind=title></h4><span class=total-bar><span class=bar-portion ng-repeat=\"field in fields\" style=\"width: {{ field.percentage }}%\"></span> </span><span class=total>Total: <span class=total-figure ng-bind=total></span></span><div class=fields-wrapper><div class=proportion-field ng-repeat=\"field in fields\"><div drf-form-field=field class=\"proportion-field-inner {{ form[id].$invalid ? 'has-error' : '' }}\"><div class=proportion-input integer-field model=model[$index] id=field.id form=form></div><span class=percentage ng-class=\"{ filled: field.percentage > 0 }\"><span class=number ng-bind=field.percentage|number:0></span>%</span></div></div></div>"
+    "<h4 class=title ng-bind=title></h4><div class=fields-wrapper><div class=proportion-field ng-repeat=\"field in fields\"><div drf-form-field=field class=\"proportion-field-inner {{ form[id].$invalid ? 'has-error' : '' }}\">AMOUNT ALLOCATED (use for css): {{ total - model[$index] }}<div class=proportion-input integer-field model=model[$index] id=field.id form=form></div><div aif-slider-input model=model[$index] ceiling=100 slider-low-label=0% slider-high-label=100%></div></div></div></div>"
   );
 
 
   $templateCache.put('templates/incuna-surveys/form/base/survey-form.html',
-    "<form class=question-form ng-submit=submit()><div calculate-percentage question-set=fields model=model></div><header class=form-header><h1 ng-bind=form.name class=form-name></h1><h2 ng-bind=form.description class=form-desc></h2></header><formly-form class=form-body model=model fields=fields></formly-form><button class=button-submit type=submit>Submit</button></form>"
+    "<form class=question-form ng-submit=submit()><div calculate-percentage question-set=fields model=model></div><header class=form-header><h1 ng-bind=form.name class=form-name></h1><h2 ng-bind=form.description class=form-desc></h2></header><formly-form class=form-body model=model fields=fields></formly-form><button ng-show=form class=button-submit type=submit translate>Submit</button></form>"
   );
 
 
@@ -32,12 +32,12 @@ angular.module('incuna-surveys-form.templates', []).run(['$templateCache', funct
 
 
   $templateCache.put('templates/incuna-surveys/form/proportion-field.html',
-    "<h4 class=title ng-bind=title></h4><span class=total-bar><span class=bar-portion ng-repeat=\"field in fields\" style=\"width: {{ field.percentage }}%\"></span> </span><span class=total>Total: <span class=total-figure ng-bind=total></span></span><div class=fields-wrapper><div class=proportion-field ng-repeat=\"field in fields\"><div drf-form-field=field class=\"proportion-field-inner {{ form[id].$invalid ? 'has-error' : '' }}\"><div class=proportion-input integer-field model=model[$index] id=field.id form=form></div><span class=percentage ng-class=\"{ filled: field.percentage > 0 }\"><span class=number ng-bind=field.percentage|number:0></span>%</span></div></div></div>"
+    "<h4 class=title ng-bind=title></h4><div class=fields-wrapper><div class=proportion-field ng-repeat=\"field in fields\"><div drf-form-field=field class=\"proportion-field-inner {{ form[id].$invalid ? 'has-error' : '' }}\">AMOUNT ALLOCATED (use for css): {{ total - model[$index] }}<div class=proportion-input integer-field model=model[$index] id=field.id form=form></div><div aif-slider-input model=model[$index] ceiling=100 slider-low-label=0% slider-high-label=100%></div></div></div></div>"
   );
 
 
   $templateCache.put('templates/incuna-surveys/form/survey-form.html',
-    "<form class=question-form ng-submit=submit()><div calculate-percentage question-set=fields model=model></div><header class=form-header><h1 ng-bind=form.name class=form-name></h1><h2 ng-bind=form.description class=form-desc></h2></header><formly-form class=form-body model=model fields=fields></formly-form><button class=button-submit type=submit>Submit</button></form>"
+    "<form class=question-form ng-submit=submit()><div calculate-percentage question-set=fields model=model></div><header class=form-header><h1 ng-bind=form.name class=form-name></h1><h2 ng-bind=form.description class=form-desc></h2></header><formly-form class=form-body model=model fields=fields></formly-form><button ng-show=form class=button-submit type=submit translate>Submit</button></form>"
   );
 
 }]);