+++ web/app/scripts/src/services/proportion-field.es.js
@@ -34,19 +34,9 @@ module.service(moduleProperties.componentName, [
                 return 0;
             }
             return Object.keys(values).reduce((total, key) => {
-                const value = parseInt(values[key], 10);
-                if (Number.isNaN(value)) {
-                    return total;
-                }
+                const value = parseInt(values[key], 10) || 0;
                 return total + value;
             }, 0);
-        };
-
-        this.addPercentages = function (fields, values, total) {
-            fields.forEach((options, key) => {
-                const value = parseInt(values[key], 10) || 0;
-                options.percentage = value ? value / total * 100 : 0;
-            });
         };
 
         this.addErrors = function (fields, errors) {
@@ -54,6 +44,13 @@ module.service(moduleProperties.componentName, [
                 options.errors = errors[index];
             });
         };
+
+        this.getChangedFieldIndex = function (newModel, oldModel) {
+            const newValues = Object.values(newModel);
+            const oldValues = Object.values(oldModel);
+
+            return newValues.findIndex((el, i) => el !== oldValues[i]);
+        }
 
     }
 ]);