+++ changelog.md
@@ -2,6 +2,10 @@
 
 This project uses Semantic Versioning (2.0).
 
+# upcoming
+
+* Change proportion field to a slider group
+
 ### 0.14.0
 
 * The response create serializer will now correctly return the instance created